Combinators replace existing properties.
In general, it makes no sense to add the same property to an
option multiple times. When it does make sense, it's probably not
very useful. The + operation on option properties is commutative
in some cases (e.g. names), but not in others (like help strings,
which must be concatenated in some order).

For now, property manipulation is split between some important
optional arguments to 'option' and the opt- combinators. This is
a little awkward, but, for now, I prefer it to the "constructor
with 18 named, optional parameters" approach.